requirements.txt
fastapi
uvicorn
python-dotenv
requests
schedule
python-dotenv